+++ core/wallet-auth/src/auth-utils.test.ts
@@ -17,24 +17,11 @@ import {
     fetchOidcUserInfo,
     jwtExpired,
 } from './auth-utils.js'
-import {
-    AuthTokenProvider,
-    TokenProviderConfig,
-} from './auth-token-provider.js'
+import { TokenProviderConfig } from './auth-token-provider.js'
 import { Logger } from '@canton-network/core-types'
-import { Auth, Idp } from './config/schema.js'
-import { clientCredentialsService } from './client-credentials-service.js'
 import { SelfSignedTokenService } from './self-signed-token-service.js'
 
-vi.mock('./client-credentials-service.js', () => {
-    return {
-        clientCredentialsService: vi.fn(() => ({
-            fetchToken: vi.fn(),
-        })),
-    }
-})
-
-describe('AuthTokenProvider', () => {
+describe('Auth Utils', () => {
     beforeEach(() => {
         const fetchMock = vi.fn()
 
@@ -62,97 +49,6 @@ describe('AuthTokenProvider', () => {
         warn: vi.fn(),
         error: vi.fn(),
     } as MockedObject<Logger>
-
-    it('should test an auth token provider initialization from TokenProviderConfig', async () => {
-        const authProvider = new AuthTokenProvider(
-            tokenProviderConfig,
-            mockLogger
-        )
-
-        const authContext = await authProvider.getAuthContext()
-        expect(assertConnected(authContext)).toBe(authContext)
-        expect(jwtExpired(authContext.accessToken)).toBeFalsy()
-    })
-
-    it('should test an auth token provider initialization with fromToken', async () => {
-        const token = await SelfSignedTokenService.fetchToken(
-            mockLogger,
-            tokenProviderConfig.credentials,
-            tokenProviderConfig.issuer
-        )
-
-        const authProviderFromToken = AuthTokenProvider.fromToken(
-            token,
-            mockLogger
-        )
-        const authContext = await authProviderFromToken.getAuthContext()
-        expect(assertConnected(authContext)).toBe(authContext)
-    })
-
-    it('should test an auth token provider initialization with fromGatewayConfig', async () => {
-        const idp: Idp = {
-            id: 'test1',
-            type: 'self_signed',
-            issuer: 'unsafe-auth',
-        }
-
-        const auth: Auth = {
-            method: 'self_signed',
-            issuer: 'unsafe-auth',
-            clientId: 'ledger-api-user',
-            clientSecret: 'unsafe',
-            audience: 'https://canton.network.global',
-            scope: '',
-        }
-
-        const authProviderFromGateway = AuthTokenProvider.fromGatewayConfig(
-            idp,
-            auth,
-            mockLogger
-        )
-
-        const authContext = await authProviderFromGateway.getAuthContext()
-        expect(assertConnected(authContext)).toBe(authContext)
-    })
-
-    it('should test an auth token provider initialization with fromGatewayConfig with oauth', async () => {
-        const authProvider = new AuthTokenProvider(
-            tokenProviderConfig,
-            mockLogger
-        )
-
-        const token = await authProvider.getAccessToken()
-
-        const idp: Idp = {
-            id: 'test1',
-            type: 'oauth',
-            issuer: 'unsafe-auth',
-            configUrl,
-        }
-
-        const auth: Auth = {
-            method: 'client_credentials',
-            clientId: 'ledger-api-user',
-            clientSecret: 'unsafe',
-            audience: 'https://canton.network.global',
-            scope: '',
-        }
-
-        const mockedServiceFactory = vi.mocked(clientCredentialsService)
-
-        const mockFetchToken = vi.fn().mockResolvedValue(token)
-
-        mockedServiceFactory.mockReturnValue({ fetchToken: mockFetchToken })
-
-        const authProviderFromGateway = AuthTokenProvider.fromGatewayConfig(
-            idp,
-            auth,
-            mockLogger
-        )
-
-        const authContext = await authProviderFromGateway.getAuthContext()
-        expect(assertConnected(authContext)).toBe(authContext)
-    })
 
     it('should verify components of a jwt token correctly', async () => {
         const token = await SelfSignedTokenService.fetchToken(
